News_data/requirements.txt
requests
pandas
yfinance
python-dotenv 